requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.3
contourpy==1.0.6
cycler==0.11.0
datasets==2.8.0
dill==0.3.6
EasyNMT==2.0.2
faiss-gpu==1.7.2
fasttext==0.9.2
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.24.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.3
MarkupSafe==2.1.2
matplotlib==3.6.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
nltk==3.8.1
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.6
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pyarrow==10.0.1
pycocoevalcap==1.2
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.5.4
sentencepiece==0.1.97
simcse==0.4
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2022.10.10
tiktoken==0.4.0
timm==0.4.9
tokenizers==0.10.3
torch==2.0.1
tqdm==4.64.1
transformers==4.12.5
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
xxhash==3.2.0
yarl==1.8.2
